requirements.txt
customtkinter
pillow
flask
